ATF 3.10. See ReleaseNotes.doc.
This release reflects six months of bug fixes and minor features. ATF is
now in maintenance mode, with only bug fixes and minor client-requested
features being put in by Sony Computer Entertainment’s Worldwide
Studios. Our open source users are encouraged to contribute on GitHub.

+++ Framework/Atf.Gui/Gdi32.cs
@@ -10,10 +10,10 @@ namespace Sce.Atf
     [StructLayout(LayoutKind.Sequential)]
     public struct BITMAP
     {
-        public long bmType;
-        public long bmWidth;
-        public long bmHeight;
-        public long bmWidthBytes;
+        public int bmType;
+        public int bmWidth;
+        public int bmHeight;
+        public int bmWidthBytes;
         public short bmPlanes;
         public short bmBitsPixel;
         public IntPtr bmBits;
